requirements.txt
discord.py
python-aternos
dotenv